+++ primefaces/src/main/java/org/primefaces/component/autocomplete/AutoComplete.java
@@ -25,6 +25,7 @@ package org.primefaces.component.autocomplete;
 
 import java.util.*;
 import javax.el.MethodExpression;
+import javax.faces.FacesException;
 import javax.faces.application.ResourceDependency;
 import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
@@ -34,6 +35,9 @@ import org.primefaces.component.api.AbstractPrimeHtmlInputText;
 import org.primefaces.component.column.Column;
 import org.primefaces.event.SelectEvent;
 import org.primefaces.event.UnselectEvent;
+import org.primefaces.model.FilterMeta;
+import org.primefaces.model.LazyDataModel;
+import org.primefaces.model.LazySearchable;
 import org.primefaces.util.ComponentUtils;
 import org.primefaces.util.Constants;
 import org.primefaces.util.LangUtils;
@@ -72,6 +76,8 @@ public class AutoComplete extends AutoCompleteBase {
     protected static final Collection<String> EVENT_NAMES = LangUtils.concat(AbstractPrimeHtmlInputText.EVENT_NAMES, UNOBSTRUSIVE_EVENT_NAMES);
 
     private Object suggestions;
+
+    private Integer suggestionsCount;
 
     @Override
     public Collection<String> getEventNames() {
@@ -135,18 +141,39 @@ public class AutoComplete extends AutoCompleteBase {
     public void broadcast(javax.faces.event.FacesEvent event) throws javax.faces.event.AbortProcessingException {
         super.broadcast(event);
 
-        FacesContext facesContext = getFacesContext();
-        MethodExpression me = getCompleteMethod();
-
-        if (me != null && event instanceof org.primefaces.event.AutoCompleteEvent) {
-            suggestions = me.invoke(facesContext.getELContext(), new Object[]{((org.primefaces.event.AutoCompleteEvent) event).getQuery()});
-
-            if (suggestions == null) {
-                suggestions = isServerQueryMode() ? new ArrayList() : new HashMap<String, List<String>>();
-            }
-
-            facesContext.renderResponse();
+        if (!(event instanceof org.primefaces.event.AutoCompleteEvent)) {
+            return;
         }
+
+        String query = ((org.primefaces.event.AutoCompleteEvent) event).getQuery();
+        LazyDataModel lazyModel = getLazyModel();
+        if (lazyModel != null) {
+            if (!(lazyModel instanceof LazySearchable)) {
+                throw new FacesException("lazyModel should implement LazySearchable");
+            }
+            Map<String, FilterMeta> searchFilter = ((LazySearchable) lazyModel).getSearchFilter(query);
+            suggestions = lazyModel.load(0, getMaxResults(), Collections.emptyMap(), searchFilter);
+            suggestionsCount = lazyModel.count(searchFilter);
+        }
+        else {
+            FacesContext facesContext = getFacesContext();
+            MethodExpression me = getCompleteMethod();
+
+            if (me != null) {
+                suggestions = me.invoke(facesContext.getELContext(), new Object[]{query});
+
+                if (suggestions == null) {
+                    suggestions = isServerQueryMode() ? Collections.emptyList() : Collections.emptyMap();
+                }
+
+                facesContext.renderResponse();
+            }
+        }
+    }
+
+    protected boolean hasMoreSuggestions() {
+        int count = suggestionsCount != null ? suggestionsCount : ((List) getSuggestions()).size();
+        return count > getMaxResults();
     }
 
     public List<Column> getColums() {
@@ -202,6 +229,7 @@ public class AutoComplete extends AutoCompleteBase {
     public Object saveState(FacesContext context) {
         // reset component for MyFaces view pooling
         suggestions = null;
+        suggestionsCount = null;
 
         return super.saveState(context);
     }