+++ primefaces/src/main/java/org/primefaces/component/autocomplete/AutoCompleteBase.java
@@ -27,6 +27,7 @@ import org.primefaces.component.api.AbstractPrimeHtmlInputText;
 import org.primefaces.component.api.InputHolder;
 import org.primefaces.component.api.MixedClientBehaviorHolder;
 import org.primefaces.component.api.Widget;
+import org.primefaces.model.LazyDataModel;
 import org.primefaces.util.MessageFactory;
 
 public abstract class AutoCompleteBase extends AbstractPrimeHtmlInputText implements Widget, InputHolder, MixedClientBehaviorHolder {
@@ -82,7 +83,8 @@ public abstract class AutoCompleteBase extends AbstractPrimeHtmlInputText implem
         queryMode,
         dropdownTabindex,
         dropdownAriaLabel,
-        completeEndpoint
+        completeEndpoint,
+        lazyModel
     }
 
     public AutoCompleteBase() {
@@ -445,4 +447,13 @@ public abstract class AutoCompleteBase extends AbstractPrimeHtmlInputText implem
     public void setCompleteEndpoint(String completeEndpoint) {
         getStateHelper().put(PropertyKeys.completeEndpoint, completeEndpoint);
     }
+
+    public LazyDataModel getLazyModel() {
+        return (LazyDataModel) getStateHelper().eval(PropertyKeys.lazyModel, null);
+    }
+
+    public void setCompleteEndpoint(LazyDataModel lazyModel) {
+        getStateHelper().put(PropertyKeys.lazyModel, lazyModel);
+    }
+
 }