+++ docs/12_0_0/gettingstarted/whatsnew.md
@@ -4,5 +4,6 @@ This page contains a list of big features. Please check the GitHub issues for al
 
 ## 12.0.0
 
+  * AutoComplete: added `lazyModel` attribute for fetching autocomplete-suggestions (instead of completeMethod).
 
 Look into [migration guide](https://primefaces.github.io/primefaces/12_0_0/#/../migrationguide/12_0_0) for more enhancements and changes.